will not produce .j file when error occurs

+++ test.c
@@ -2,11 +2,11 @@ void main(){
     int a = 10;
     int b = 10;
 
-    float c= 100;
+    int c= 100;
     while(a > 0){
         c %= a;
         print(c);
-        a--; 
+        a--;
     }
     return;
 }
